refactor: split service_crud_test.go from project/service_test.go
